refactor(core): break api↔executor and executor↔optimizer cycles
Move four shared items into `core` so executor and optimizer can
name them without reaching back through the api facade. Both
cross-module cycles in the dep graph collapse to a strict DAG,
which is a prerequisite for splitting the crate into a Cargo
workspace later.

Moved:
  - ParamVec (SmallVec<[Value; 8]>): from api::params to
    core::mod. api::params now re-exports for path stability;
    public `stoolap::ParamVec` and `crate::api::params::ParamVec`
    both still work.
  - DEFAULT_PARALLEL_FILTER_THRESHOLD,
    DEFAULT_PARALLEL_SORT_THRESHOLD,
    DEFAULT_PARALLEL_JOIN_THRESHOLD: from executor::parallel to
    core::mod. executor::parallel now re-exports for path
    stability; DEFAULT_PARALLEL_CHUNK_SIZE stays in executor
    (executor-only).

Import flips at the actual cycle edges (so `cargo metadata` sees
a clean DAG, not just a re-export hop):
  - executor/mod.rs, executor/context.rs,
    executor/expression/evaluator_bridge.rs: import ParamVec
    from core directly (was api::params).
  - optimizer/cost.rs: import the three thresholds from core
    directly (was executor).

Verification:
  - `grep -r 'use crate::api' src/executor/` -> empty
  - `grep -r 'use crate::executor' src/optimizer/` -> empty
  - cargo fmt clean
  - cargo clippy --all-targets --all-features -D warnings clean
  - cargo nextest run (lib + tests): 5082 passed, 2 skipped

Net diff: +14 / -16 LOC across 7 files. No behavior change.
